[redhat] New configs in mm/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 Symbol: PAGE_BLOCK_ORDER [=10]
 Type  : integer
 Range : [1 10]
 Defined at mm/Kconfig:1008
   Prompt: Page Block Order
   Location:
     -> Memory Management options
       -> Page Block Order (PAGE_BLOCK_ORDER [=10])

 Commit: e13e7922 (mm: add CONFIG_PAGE_BLOCK_ORDER to select page block order)

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>