Set CONFIG_TEST_VMALLOC to off for s390 zfcpdump
Upstream removed the dep on module loading from CONFIG_TEST_VMALLOC with
commit 2d76e79315e403aab595d4c8830b7a46c19f0f3b which created a mismatch
for the zfcpdump kernel. RHEL sets CONFIG_TEST_VMALLOC=m but now that
module support is not required, we must either set it to built in or off
for s390 zfcpdump. Given that it was off prior to 6.16 kernes because it
did not meet the dep requirements, let's specifically turn it off.

Signed-off-by: Justin M. Forbes <jforbes@fedoraproject.org>