[redhat] New configs in drivers/pwm
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: PWM_ARGON_FAN_HAT [=n]
 Type  : tristate
 Defined at drivers/pwm/Kconfig:69
   Prompt: Argon40 Fan HAT support
   Depends on: PWM [=y] && I2C [=y] && OF [=y]
   Location:
     -> Device Drivers
       -> Pulse-Width Modulation (PWM) Support (PWM [=y])
         -> Argon40 Fan HAT support (PWM_ARGON_FAN_HAT [=n])

 ~~~
 Commit: 0191c80e8a28 (pwm: argon-fan-hat: Add Argon40 Fan HAT support)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>