[redhat] New configs in init/Kconfig
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: IO_URING_MOCK_FILE [=n]
 Type  : tristate
 Defined at init/Kconfig:1808
   Prompt: Enable io_uring mock files (Experimental)
   Depends on: IO_URING [=y]
   Visible if: IO_URING [=y] && EXPERT [=y]
   Location:
     -> General setup
       -> Configure standard kernel features (expert users) (EXPERT [=y])
         -> Enable IO uring support (IO_URING [=y])
           -> Enable io_uring mock files (Experimental) (IO_URING_MOCK_FILE [=n])

 ~~~
 Commit: 3a0ae385f69e (io_uring/mock: add basic infra for test mock files)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>