[redhat] New configs in arch/x86
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: KVM_IOAPIC [=y]
 Type  : bool
 Defined at arch/x86/kvm/Kconfig:169
   Prompt: I/O APIC, PIC, and PIT emulation
   Depends on: VIRTUALIZATION [=y] && KVM [=m]
   Location:
     -> Virtualization (VIRTUALIZATION [=y])
       -> I/O APIC, PIC, and PIT emulation (KVM_IOAPIC [=y])

 ~~~
 Commit: 628a27731e3f (KVM: x86: Add CONFIG_KVM_IOAPIC to allow disabling in-kernel I/O APIC)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>