[redhat] New configs in kernel/trace
Hi,

As part of the ongoing rebase effort, the following configuration
options need to be reviewed.

As a reminder, the ARK configuration flow involves moving unreviewed
configuration options from the pending directory to the ark directory.
In the diff below, options are removed from the pending directory and
added to the ark hierarchy. The final options that need to be ACKed
are the files that are being added to the ark hierarchy.

If the value for a file that is added should be changed, please reply
with a better option.

 ~~~
 Symbol: RV_MON_RTAPP [=n]
 Type  : bool
 Defined at kernel/trace/rv/monitors/rtapp/Kconfig:1
   Prompt: rtapp monitor
   Depends on: FTRACE [=y] && RV [=y] && RV_PER_TASK_MONITORS [=2]>=2
   Location:
     -> Kernel hacking
       -> Tracers (FTRACE [=y])
         -> Runtime Verification (RV [=y])
           -> rtapp monitor (RV_MON_RTAPP [=n])

 ~~~
 Commit: 886fc86e9419 (rv: Add rtapp container monitor)

---

 ~~~
 Symbol: RV_PER_TASK_MONITORS [=2]
 Type  : integer
 Range : [1 8]
 Defined at kernel/trace/rv/Kconfig:40
   Prompt: Maximum number of per-task monitor
   Depends on: FTRACE [=y] && RV [=y]
   Location:
     -> Kernel hacking
       -> Tracers (FTRACE [=y])
         -> Runtime Verification (RV [=y])
           -> Maximum number of per-task monitor (RV_PER_TASK_MONITORS [=2])

 ~~~
 Commit: fac5493251a6 (rv: Allow to configure the number of per-task monitor)

---

Signed-off-by: Fedora Kernel Team <kernel-team@fedoraproject.org>