Input: gpio_keys: Temporarily make KEY_POWER instant
Systemd has a bug where it will process the KEY_POWER event that woke
the system after it woke up, if it was produced by gpio_keys (it supports
holding the button). This will make the system sleep instantly again and
crash it.

Therefore, mute the event on wake-up checks. Power_key should only be reported when driven as an interrupt.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>