redhat: Temporary stop adding 'kernel' component to SBAT
Two events happened recently:
- dracut-107 switched to using 'ukify' for building the UKI
- Fedora/ARK enabled CONFIG_EFI_SBAT_FILE

Unfortunately, the limit for SBAT in systemd-stub is 512 bytes so 'ukify'
breaks with

 __main__.PEError: Not enough space in existing section .sbat to append new data.

While the problem is hidden in the builds now as kernel-uki-virt does
not add the required SBAT data at all, to fix it, we need to shrink the
overall size to not hit the limit. Remove the future-proof 'kernel'
component while leaving 'kernel.<suffix>': this should be enough if
revocations are needed. The unbranded 'kernel' is supposed to simplify
revocations for the situations when all kernels below certain versions are
considered unsafe but as Fedora/CentOS,RHEL/... use different signing keys,
the SBAT revocation will need to go to different shims.

The current plan is to restore 'kernel' SBAT component when systemd-stub is
fixed. The work is being done upstream:

https://github.com/systemd/systemd/pull/38621

Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>