redhat: Explicitly disable 'hostonly' mode on the dracut cmdline
Despite the fact that we have hostonly="no" and also override
/etc/dracut.conf.d directory with --confdir=$(mktemp -d), 'hostonly' mode
still gets enabled in the builds. This can be seen in the logs:

  00:30:55 dracut[W]: Running in hostonly mode in a container!

This is undesired. In particular, 'hostonly' mode tries to grab /etc/shadow
from the system and obviously fails. This leads to broken emergency
shell in the UKI's initramfs. Other issues are also possible as in
'hostonly' mode the result depends a lot on the building environment.
The reason for such behavior is that dracut has a *third' place where
it gets its configuration (/usr/lib/dracut/dracut.conf.d/) and this
one can't be overriden.

Make sure we always disable 'hostonly' by adding 'no-hostonly' cmdline
switch: this gets priority over any configuration file.

Signed-off-by: Vitaly Kuznetsov <vkuznets@redhat.com>