Add 1010-config-newlines-test.bats self test.
The test checks  CONFIG_* files that don't end with
a newline and generate an error otherwise.

In build_configs.sh, the CONFIG_* files in redhat/configs
are concatenated together in combine_config_layer() to
create the initial config but if a CONFIG file doesn't end
with a newline this can cause the setting to go unrecognized.

Signed-off-by: Alexandra Hájková <ahajkova@redhat.com>