ALSA: hda/tas2781: fix speaker id retrieval for multiple probes
Currently, on ASUS projects, the TAS2781 codec attaches the speaker
GPIO to the first tasdevice_priv instance using devm. This causes
tas2781_read_acpi to fail on subsequent probes since the GPIO
is already managed by the first device. But this is unnecessary as
the GPIO contains a static value.

Therefore, instead of attaching the GPIO to the device and reading it
when loading the firmware, read its value directly in tas2781_read_acpi
and store it in the private data structure. Then, make reading the value
non-fatal so that ASUS projects that miss a speaker pin can still work,
perhaps using fallback firmware.

Fixes: 4e7035a75da9 ("ALSA: hda/tas2781: Add speaker id check for ASUS projects")
Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>