Input: gpio_keys - handle KEY_POWER gracefully during resume
Currently, during resume from suspend, if an ISR runs for a gpio-key
the ISR will rightfully emit a key press event so that userspace does
not miss the event (e.g., on an embedded device we might sleep until
a user presses a button; and when we wake up we need to be able to
identify the specific button).

This creates an issue when the gpio-key is KEY_POWER. For ISA/ACPI
power buttons, when the resume is caused by the power button, the power
button device increases its wake up count but it does not re-emit
KEY_POWER during resume. Therefore, both the kernel and userspace
act unexpectedly if the gpio-key driver emits KEY_POWER during resume,
by e.g., getting stuck in a suspend loop.

This is a problem in newer ASUS devices (ASUS Z13, Xbox Ally, Xbox Ally
X) as they wire the power button to a gpio pin. Therefore, specifically
for KEY_POWER gpio-keys, suppress the event during resume to mirror this
behavior. Do not do it for all keys, as certain embedded devices might
be relying on this behavior.

Signed-off-by: Antheas Kapenekakis <lkml@antheas.dev>